requirements.txt
# Core dependencies
streamlit>=1.28.0
torch>=2.0.0
transformers>=4.37.0
accelerate>=0.25.0

# Embeddings and vector store
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Document processing
PyMuPDF>=1.23.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Utilities
numpy>=1.24.0
